Rename MogTransduce.[mh] to MogTransformation.[mh]
